+++ src/lib/legacy/Zikula/Controller/AbstractAuthentication.php
@@ -12,6 +12,7 @@
  * Please see the NOTICE file distributed with this source code for further
  * information regarding copyright and licensing.
  */
+use Symfony\Component\Debug\Exception\FatalErrorException;
 
 /**
  * Abstract authentication module controller.
@@ -46,6 +47,11 @@ abstract class Zikula_Controller_AbstractAuthentication extends Zikula_AbstractC
         return ModUtil::apiFunc($this->name, 'Authentication', 'isEnabledForAuthentication', array('method' => $methodName), 'Zikula_Api_AbstractAuthentication');
     }
 
+    protected function getAuthenticationMethod($methodName)
+    {
+        return ModUtil::apiFunc($this->name, 'Authentication', 'getAuthenticationMethod', array('method' => $methodName), 'Zikula_Api_AbstractAuthentication');
+    }
+
     /**
      * Render and return the portion of the HTML log-in form containing the fields needed by this authentication module in order to log in.
      *
@@ -73,11 +79,17 @@ abstract class Zikula_Controller_AbstractAuthentication extends Zikula_AbstractC
     }
 
     /**
-     * Render and return an authentication method selector for the login page form or login block form.
+     * Renders the template that displays the authentication module's icon in the Users module's login block.
+     *
+     * Parameters passed in the $args array:
+     * -------------------------------------
+     * - string $args['form_type']   An indicator of the type of form on which the selector will appear.
+     * - string $args['form_action'] The URL to which the selector form should submit.
+     * - string $args['method']      The authentication method for which a selector should be returned.
      *
      * @param array $args All parameters passed to this function.
      *
-     * @todo In 1.4.0, throw exception here and do not allow the "getAuthenticationMethodSelector" method.
+     * @todo In 1.4.0, move content from "getAuthenticationMethodSelector" method to here.
      *
      * @return string The rendered authentication method selector for the login page or block.
      */
@@ -95,7 +107,63 @@ abstract class Zikula_Controller_AbstractAuthentication extends Zikula_AbstractC
      */
     public function getAuthenticationMethodSelector(array $args)
     {
-        throw new \LogicException('This method must be overridden in concrete class');
+        // Parameter extraction and error checking
+        if (!isset($args) || !is_array($args)) {
+            throw new \InvalidArgumentException($this->__('An invalid \'$args\' parameter was received.'));
+        }
+
+        if (!isset($args['form_type']) || !is_string($args['form_type'])) {
+            throw new \InvalidArgumentException($this->__f('An invalid form type (\'%1$s\') was received.', array(
+                    isset($args['form_type']) ? $args['form_type'] : 'NULL'))
+            );
+        }
+
+        if (!isset($args['form_action']) || !is_string($args['form_action'])) {
+            throw new \InvalidArgumentException($this->__f('An invalid form action (\'%1$s\') was received.', array(
+                    isset($args['form_action']) ? $args['form_action'] : 'NULL'))
+            );
+        }
+
+        if (!isset($args['method']) || !is_string($args['method']) || !$this->supportsAuthenticationMethod($args['method'])) {
+            throw new \InvalidArgumentException($this->__f('Error: An invalid method (\'%1$s\') was received.', array(
+                    isset($args['method']) ? $args['method'] : 'NULL'))
+            );
+        }
+        // End parameter extraction and error checking
+
+        if ($this->authenticationMethodIsEnabled($args['method'])) {
+            $authenticationMethod = $this->getAuthenticationMethod($args['method']);
+            $icon = $authenticationMethod->getIcon();
+            $isFontAwesomeIcon = $authenticationMethod->isFontAwesomeIcon();
+
+            $templateVars = array(
+                'authentication_method' => array(
+                    'modname'   => $this->name,
+                    'method'    => $args['method'],
+                ),
+                'is_selected'           => isset($args['is_selected']) && $args['is_selected'],
+                'form_type'             => $args['form_type'],
+                'form_action'           => $args['form_action'],
+                'submit_text'           => $authenticationMethod->getShortDescription(),
+                'icon'                  => $icon,
+                'isFontAwesomeIcon'     => $isFontAwesomeIcon,
+            );
+
+            $view = new Zikula_View($this->serviceManager, 'ZikulaUsersModule', Zikula_View::CACHE_ENABLED);
+            $view->assign($templateVars);
+
+            $templateName = "Authentication/AuthenticationMethodSelector/{$args['form_type']}/Base.tpl";
+            if (!$view->template_exists($templateName)) {
+                $templateName = "Authentication/AuthenticationMethodSelector/Base/Base.tpl";
+                if (!$view->template_exists($templateName)) {
+                    throw new FatalErrorException($this->__f('A form fields template was not found for the %1$s method using form type \'%2$s\'.', array($args['method'], $args['form_type'])));
+                }
+            }
+
+            return $this->response($view->fetch($templateName));
+        } else {
+            return $this->response('');
+        }
     }
 
     /**